+++ storage/src/test/java/ua/com/tracksee/dao/postgresql/TaxiOrderDAOBeanTest.java
@@ -64,7 +64,7 @@ public class TaxiOrderDAOBeanTest {
         order.setFreeWifi(TRUE);
         order.setDescription("I like to mov it mov it");
 
-        Long trackingNumber = taxiOrderDAO.addOrder(order);
+        Integer trackingNumber = taxiOrderDAO.addOrder(order);
 
         TaxiOrderEntity databaseOrder = taxiOrderDAO.getOrder(trackingNumber);
         assertEquals(databaseOrder.getTrackingNumber(), trackingNumber);
